// src/types/producers.types.ts
export type TipoPersona = "FISICA" | "JURIDICA";

export type EstadoProductora = "Autorizada" | "Pendiente";

export const TIPO_DOCUMENTOS = [
  "dni_persona_fisica",
  "contrato_social",
  "dni_representante_legal",
  "comprobante_ISRC",
] as const;

export type TipoDocumento = (typeof TIPO_DOCUMENTOS)[number];